src/requirements.txt
fastapi==0.123.5
uvicorn[standard]==0.38.0
pillow==12.0.0
pydantic==2.12.5
python-multipart==0.0.20
prometheus-client==0.21.0
redis==7.1.0
starlette==0.50.0
